Move tests to keycloak-tests-base
